+++ test/helper.js
@@ -161,7 +161,7 @@ module.exports = {
     },
     killConnection: function (client) {
         // Change the connection option to a non existing one and destroy the stream
-        client.connectionOption = {
+        client.connection_option = {
             port: 65535,
             host: '127.0.0.1',
             family: 4